Refactored entire code base
- Moved functions and other into different files
- Simplified some code
- Used `cargo clippy` to improve code
